+++ composeApp/src/jvmMain/kotlin/ru/atrs/mcm/parsing_excel/models/PressuresHolder.kt
@@ -3,8 +3,8 @@ package ru.atrs.mcm.parsing_excel.models
 data class PressuresHolder(
     val displayName : String,
     val index : Int,
-    var minValue : Int,
-    var maxValue : Int,
+    var minValue : Float,
+    var maxValue : Float,
     val tolerance : Int,
     val unit : String,
     val commentString : String,
